+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/modals/constants.ts
@@ -1,4 +1,3 @@
 export * from './publish-with-descendants-modal/constants.js';
 export * from './save-modal/constants.js';
-export * from './unpublish-modal/constants.js';
 export * from './document-picker-modal.token.js';
