+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/modals/manifests.ts
@@ -1,9 +1,4 @@
 import { manifest as publishWithDescendantsModalManifest } from './publish-with-descendants-modal/manifest.js';
 import { manifest as saveModalManifest } from './save-modal/manifest.js';
-import { manifest as unpublishModalManifest } from './unpublish-modal/manifest.js';
 
-export const manifests: Array<UmbExtensionManifest> = [
-	publishWithDescendantsModalManifest,
-	saveModalManifest,
-	unpublishModalManifest,
-];
+export const manifests: Array<UmbExtensionManifest> = [publishWithDescendantsModalManifest, saveModalManifest];
