--- src/Umbraco.Web.UI.Client/src/packages/documents/documents/modals/unpublish-modal/constants.ts
@@ -1,2 +1,0 @@
-export * from './document-unpublish-modal.token.js';
-export { UMB_DOCUMENT_UNPUBLISH_MODAL_ALIAS } from './manifest.js';
