+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/publishing/manifests.ts
@@ -1,6 +1,7 @@
 import { manifests as publishManifest } from './publish/manifests.js';
 import { manifests as repositoryManifests } from './repository/manifests.js';
 import { manifests as schedulePublishManifests } from './schedule-publish/manifests.js';
+import { manifests as unpublishManifests } from './unpublish/manifests.js';
 import { UMB_DOCUMENT_WORKSPACE_ALIAS } from '../workspace/constants.js';
 import { UMB_WORKSPACE_CONDITION_ALIAS } from '@umbraco-cms/backoffice/workspace';
 
@@ -20,4 +21,5 @@ export const manifests: Array<UmbExtensionManifest> = [
 	...publishManifest,
 	...repositoryManifests,
 	...schedulePublishManifests,
+	...unpublishManifests,
 ];
