+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/publishing/types.ts
@@ -1,2 +1,3 @@
 export type * from './publish/types.js';
 export type * from './schedule-publish/types.js';
+export type * from './unpublish/types.js';
