+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/publishing/unpublish/manifests.ts
@@ -1,0 +1,3 @@
+import { manifests as modalManifests } from './modal/manifests.js';
+
+export const manifests: Array<UmbExtensionManifest> = [...modalManifests];
