+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/publishing/unpublish/modal/document-unpublish-modal.stories.ts
@@ -1,6 +1,6 @@
 import './document-unpublish-modal.element.js';
 
-import { UmbDocumentVariantState } from '../../types.js';
+import { UmbDocumentVariantState } from '../../../types.js';
 import type {
 	UmbDocumentUnpublishModalData,
 	UmbDocumentUnpublishModalValue,
