+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/publishing/unpublish/modal/manifests.ts
@@ -1,0 +1,10 @@
+import { UMB_DOCUMENT_UNPUBLISH_MODAL_ALIAS } from './constants.js';
+
+export const manifests: Array<UmbExtensionManifest> = [
+	{
+		type: 'modal',
+		alias: UMB_DOCUMENT_UNPUBLISH_MODAL_ALIAS,
+		name: 'Document Unpublish Modal',
+		element: () => import('./document-unpublish-modal.element.js'),
+	},
+];
