+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/publishing/unpublish/types.ts
@@ -1,0 +1,1 @@
+export type * from './modal/types.js';
